requirements.txt
streamlit==1.34.0
pandas==2.1.1
python-dotenv>=1.0.0
pytest>=7.4.0
boto3==1.34.14
toml>=0.10.2
Pillow==10.1.0
openpyxl==3.1.2  # For Excel file handling 